Create CODE_OF_CONDUCT.md
This is based on the contributer covenant, the gold standard for COCs at the moment.

At some point we should have a conversation on the specifics of enforcing violations and setting up processes for it (which will be benificial for our other open source projects too) , but that's for another day.