implement CLI- app, new, test, build, map
This commit adds an implementation for the key cli commands.

modular app <name>: creates a new repository <name>, with yarn's
workspaces enabled. It then creates a package 'app', as a default
create-react-app installation. This is then modified to work with yarn's
workspaces. We rewrite the app's webpack config to read from a widgets
folder, where we'll host other components. We also copy an
implementation for a faux 'store' to save configurations, tuples of
[component, props].

modular new <name>: This creates a new 'widget' under widgets/<name>.

modular test: runs all tests in the repository.

modular build: runs a full build, output in app/build

modular map: generates a mapping from widget names -> React.lazy() components
for those apps. This will be used in conjunction with the store to
dynamically load components as needed.

This also sets up eslint and prettier in the repo, and can be called
with yarn lint and parn prettier, respectively.