Let a user create different kinds of packages in a modular repo. (#73)
* Let a user create different kinds of packages in a modular repo.

This PR adds the ability for a user to add different types of packages to a modular repo. It does this in 4 commits: 

1. `modular build` / `modular start` accept an argument, the name of the app to build/start. It changes the included scripts in package.json to reflect this; `modular build app` and `modular start app`.

2. Next, it relocates module-template-widget-typescript into modular-scripts as templates/widget. Further, it makes templates/package for simpler packages. This commit also cleans up builds after every run. 

3. `Running modular add <name>` pops up a list of choices to select for the type of package to create; `app`, `widget` and `package`. For `widget` and `package`, it copies a folder from inside modular-scripts/templates into the desired directoy. 

4. In the final commit, selecting `app` in the list creates a new create-react-app instance in the desired location. 

followup todo: Does `modular test` only run the tests in `packages/app`? We should verify that it runs all the tests in the repo (If it does, nothing else to do here.)

addresses https://github.com/jpmorganchase/modular/issues/67, https://github.com/jpmorganchase/modular/issues/39