[e2e tests] generate ascii trees of folder structures for snapshots (#83)
This commit generates an ascii tree representation of a directory, including hashes for individual file contents. It's better than our previous implementation because it's terser (but with more info), will error when the file contents change, and is visually appealing and can be included inline in tests. Example: 

<img width="436" alt="Screenshot 2020-09-06 at 15 08 51" src="https://user-images.githubusercontent.com/18808/92327641-f5338400-f052-11ea-9e3a-eeceba88f322.png">


This commit also updates some snapshot tests with the newer published versions of packages. 

Closes https://github.com/jpmorganchase/modular/issues/52