Don't create the shared folder in a new project (#124)
* don't create the shared folder in a new project

'shared' was created in the initial iteration of modular because every other folder was a widget at the time. However, we now support creating regular packages at any time. We kept shared as a convenience, but it'll encourage people to use it, whereas we want the opposite; there has to be a higher bar to start using it. Better to not create the package by default at all. Resolves https://github.com/jpmorganchase/modular/issues/38

* add a changest

Co-authored-by: Sunil Pai <sunil.pai@jpmorgan.com>