implement modular-views.macro
This implements the view map as a babel macro, replacing app/views.ts
from our previous iteration. This is written in plain javascript because
it's otherwise hard to author and test the macro. In a future revisions,
we could add jsdoc annotations
(https://www.typescriptlang.org/docs/handbook/jsdoc-supported-types.html)