modular.tsconfig
Fixes https://github.com/jpmorganchase/modular/issues/60

This creates a new package, `modular.tsconfig`, for usage with modular projects. We can iterate on this, and further improve it later. The benefit is that we can make meaningful changes and propagate them to consumers.

This PR does _not_ make changes to `create-modular-react-app` to use this just yet, because we'll have to publish this first before we can use it in tests, so I've punted on that for the next PR.

(PS: Much like jest/babel/etc configs, would've been nice if we could hide this and prevent changes altogether, but I don't think the community/ecosystem is mature enough for that yet. We'll tackle that at some point)