migrate test command to jest directly (#155)
Fix for #112 where currently we require that an app be available in the modular structure to be able to run tests. This is because we leverage craco and therefore need there to be a create-react-app package available.

This PR is the first step in migrating directly to jest for modular test. I've exposed --watchAll and --coverage for now, but we should look to work out exactly what abstraction we want to create on top of jest. This also will allow us to leverage incremental testing in future.

---

* initial cut of jest testing map

* pass file arguments to test, enable tests for 'test' in CI

* pass file arguments to test, enable tests for 'test' in CI

* much like CRA, run some boilerplate validations before running tests

* add a changeset.

* address review comments

---

Co-authored-by: Luke Sheard <luke.sheard@jpmorgan.com>
Co-authored-by: Sunil Pai <sunil.pai@jpmorgan.com>