Some basic settings for vscode (#199)
* Some basic settings for vscode

This is the start of the work to automate project settings in IDEs. This PR makes it so that every modular project, when opened in vscode, gets a recommendation to install the eslint plugin, some editor settings, and a pointer to typescript so it picks up the right version. We'll expand on this in the future.

* update snapshot, rewrite populating a project and catches a bug where setupEnvironment wasn't being copied over

* missed eslintignore

* add a couple more extensions

Co-authored-by: Sunil Pai <sunil.pai@jpmorgan.com>