move the puppeteer dependency to the root (#252)
This PR moves `puppeteer`, which was an optional dependency for `modular-scripts`, into the root `package.json` as a dependency. We'd changed it from a dev dependency to an optional dependency because it's not working on Apple M1 machines at the moment. The problem is, that meant it even installs for consumers, making the install flow slower (and occasionally failing because of the network fetch). By moving it into the root here, it won't get installed for consumers but lets us carry on development. Once the underlying issue is fixed, we'll move it back into deDependencies for `modular-scripts`.